checking out baby

+++ server/schemas/typeDefs.js
@@ -105,14 +105,14 @@ const typeDefs = gql`
       protectiveness: Int
       trainability: Int
       barking: Int
-      minHeightFemale: Int
-      maxHeightFemale: Int
-      minWeightFemale: Int
-      maxWeightFemale: Int
-      minHeightMale: Int
-      maxHeightMale: Int
-      minWeightMale: Int
-      maxWeightMale: Int
+      minHeightFemale: Float
+      maxHeightFemale: Float
+      minWeightFemale: Float
+      maxWeightFemale: Float
+      minHeightMale: Float
+      maxHeightMale: Float
+      minWeightMale: Float
+      maxWeightMale: Float
     ): Auth
   }
 `;
